Start inling rust-cuda-builder functionality into rust-cuda-kernel
