--- rust-cuda-builder/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "rust-cuda-builder"
-version = "0.1.0"
-authors = ["Denys Zariaiev <denys.zariaiev@gmail.com>", "Juniper Tyree <juniper.tyree@helsinki.fi>"]
-repository = { workspace = true }
-license = "MIT"
-edition = { workspace = true }
-rust-version = { workspace = true }
-
-[dependencies]
-anyhow = { workspace = true }
-thiserror = { workspace = true }
-colored = { workspace = true }
-toml = { workspace = true, features = ["parse"] }
-semver = { workspace = true }
-regex = { workspace = true, features = ["std", "unicode"] }
-libc = { workspace = true }
-
-[lints]
-workspace = true
