--- rust-cuda-builder/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    // No-op, required to set the `OUT_DIR` variable
-}
