+++ rust-cuda-kernel/Cargo.toml
@@ -14,14 +14,14 @@ links = "libnvptxcompiler_static"
 proc-macro = true
 
 [dependencies]
+anyhow = { workspace = true, features = ["std"] }
 cargo_metadata = { workspace = true, features = ["builder"] }
+cargo-util = { workspace = true }
 colored = { workspace = true }
-lazy_static = { workspace = true }
 proc-macro2 = { workspace = true }
 proc-macro-error2 = { workspace = true }
 quote = { workspace = true }
-regex = { workspace = true }
-rust-cuda-builder = { workspace = true }
+regex = { workspace = true, features = ["std", "unicode"] }
 seahash = { workspace = true }
 serde_json = { workspace = true }
 strip-ansi-escapes = { workspace = true }
